requirements.txt
Django
django-ckeditor
django-jazzmin
Pillow
